moved strip load test
